+++ packages/apis/src/index.ts
@@ -4,4 +4,4 @@ export { default as ArweaveApi } from './ArweaveApi';
 export { default as CeramicApi } from './CeramicApi';
 export { default as GitHubApi } from './GitHubApi';
 export { default as SolanaApi } from './SolanaApi';
-export * from './CeramicApi'
+export * from './CeramicApi';