+++ packages/cli/src/commands/tutorial/prepublish.ts
@@ -18,11 +18,12 @@ $ builderdao tutorial prepublish near-101
 Notes:
 - The prepublish workflow deals with the builderdao-config.service to generate the 
 builderdao config and lock files, also updating the hash digest of the tutorial folder.
-`
+`;
 
 export const TutorialPrepublishCommand = () => {
   const tutorialPrepublish = new commander.Command('prepublish');
-  const log = (object: any) => _log(object, tutorialPrepublish.optsWithGlobals().key);
+  const log = (object: any) =>
+    _log(object, tutorialPrepublish.optsWithGlobals().key);
 
   const client = getClient({
     kafePk: tutorialPrepublish.optsWithGlobals().kafePk,
@@ -36,7 +37,10 @@ export const TutorialPrepublishCommand = () => {
       'Tutorial slug for complete tutorial package',
     )
     .addOption(
-      new commander.Option('--force', 'Force Rewrite the lock file base on slug').default(false)
+      new commander.Option(
+        '--force',
+        'Force Rewrite the lock file base on slug',
+      ).default(false),
     )
     .addOption(
       new commander.Option('--seed <seed>', 'Ceramic Seed')
@@ -50,23 +54,24 @@ export const TutorialPrepublishCommand = () => {
     )
     .description('Perform pre-publishing tasks')
     .helpOption('-h, --help', 'Display help for command')
-    .addHelpText(
-      'after',
-      helpText,
-    )
+    .addHelpText('after', helpText)
     .action(async (learnPackageName, options) => {
       const rootFolder = learnPackageName
         ? path.join(rootTutorialFolderPath, learnPackageName)
         : process.cwd();
       const { lock } = new BuilderDaoConfig(rootFolder);
       if (options.force) {
-        const proposal = await client.getTutorialBySlug(lock.chain.get('slug').value());
+        const proposal = await client.getTutorialBySlug(
+          lock.chain.get('slug').value(),
+        );
         lock.chain.set('proposalId', proposal.id.toNumber()).value();
         lock.chain.set('creator', proposal.creator).value();
         lock.write();
       }
-      await lock.read()
-      const proposal = await client.getTutorialBySlug(lock.chain.get('slug').value());
+      await lock.read();
+      const proposal = await client.getTutorialBySlug(
+        lock.chain.get('slug').value(),
+      );
       const ceramic = new CeramicApi({
         nodeUrl: options.nodeUrl,
       });
@@ -91,9 +96,9 @@ export const TutorialPrepublishCommand = () => {
       lock.chain.set('creator', proposal.creator).value();
       await lock.write();
       await updateHashDigestOfFolder(rootFolder);
-      await lock.read()
+      await lock.read();
       log(lock.data);
     });
 
   return tutorialPrepublish;
-}
+};