+++ packages/cli/src/utils/index.ts
@@ -54,4 +54,3 @@ export const hashSumDigest = async (path: string) => {
 // eslint-disable-next-line no-promise-executor-return
 export const sleep = async (ms: number) =>
   new Promise(resolve => setTimeout(resolve, ms));
-