+++ packages/md-formatter/src/data.ts
@@ -1,33 +1,35 @@
-import _ from 'lodash'
-import path from 'path'
-import { Low, JSONFile } from 'lowdb'
+import _ from 'lodash';
+import path from 'path';
+import { Low, JSONFile } from 'lowdb';
 
 export interface SolanaDetailsRow {
-  slug: string,
-  title: string,
-  description: string,
-  tags: string[],
-  difficulty: string,
-  id: number,
-  streamId: string,
-  creator: string,
-  reviewer1: string,
-  reviewer2: string,
-  state: string,
-  date: number,
+  slug: string;
+  title: string;
+  description: string;
+  tags: string[];
+  difficulty: string;
+  id: number;
+  streamId: string;
+  creator: string;
+  reviewer1: string;
+  reviewer2: string;
+  state: string;
+  date: number;
 }
-
 
 export class LowWithLodash<T> extends Low<T> {
-  chain: _.ExpChain<this['data']> = _.chain(this).get('data')
+  chain: _.ExpChain<this['data']> = _.chain(this).get('data');
 }
 
-const solana = path.resolve(__dirname, '../src/output.json')
-export const solanaDB = new LowWithLodash(new JSONFile<SolanaDetailsRow[]>(solana))
+const solana = path.resolve(__dirname, '../src/output.json');
+export const solanaDB = new LowWithLodash(
+  new JSONFile<SolanaDetailsRow[]>(solana),
+);
 
-
-const source = path.resolve(__dirname, '../src/tutorials-dump.json')
-export const tutorialDumpDB = new LowWithLodash(new JSONFile<Tutorial[]>(source))
+const source = path.resolve(__dirname, '../src/tutorials-dump.json');
+export const tutorialDumpDB = new LowWithLodash(
+  new JSONFile<Tutorial[]>(source),
+);
 
 export interface Tutorial {
   slug: string;
@@ -42,7 +44,6 @@ export interface Tutorial {
   published_at: string;
   pages: Page[];
 }
-
 
 export interface TutorialExtended {
   slug: string;
@@ -70,7 +71,9 @@ type Page = {
   next_slug: string;
   previous_slug: string;
   page_number: number;
-}
+};
 
-const target = path.resolve(__dirname, '../src/tutorials.json')
-export const tutorialDumpExtendedDB = new LowWithLodash(new JSONFile<{ [slug: string]: TutorialExtended }>(target))
+const target = path.resolve(__dirname, '../src/tutorials.json');
+export const tutorialDumpExtendedDB = new LowWithLodash(
+  new JSONFile<{ [slug: string]: TutorialExtended }>(target),
+);