+++ packages/apis/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @builderdao/apis
 
+## 0.0.8
+
+### Patch Changes
+
+- 4e6a667c: Added full-text search and prettier lint
+- Update version of the packages
+- 4e6a667c: update algolia full-text indexing
+- Updated dependencies [4e6a667c]
+- Updated dependencies
+  - @builderdao/program-tutorial@0.0.8
+
 ## 0.0.7
 
 ### Patch Changes