+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @builderdao/tsconfig
 
+## 0.0.6
+
+### Patch Changes
+
+- Update version of the packages
+
 ## 0.0.5
 
 ### Patch Changes